% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\alias{calculate_distance_pearson_gpu}
\alias{calculate_distance_spearman_gpu}
\alias{calculate_distance_uncentered_gpu}
\alias{calculate_distance_euclidean_gpu}
\alias{calculate_distance_pearson_cpu}
\alias{calculate_distance_spearman_cpu}
\alias{calculate_distance_uncentered_cpu}
\alias{calculate_distance_euclidean_cpu}
\alias{select_distance}
\title{calculate_distance}
\usage{
calculate_distance_pearson_gpu(x)

calculate_distance_spearman_gpu(x)

calculate_distance_uncentered_gpu(x)

calculate_distance_euclidean_gpu(x)

calculate_distance_pearson_cpu(x)

calculate_distance_spearman_cpu(x)

calculate_distance_uncentered_cpu(x)

calculate_distance_euclidean_cpu(x)

select_distance(distancetype = "pearson", usegpu = TRUE)
}
\arguments{
\item{x}{an expression matrix with features as rows and samples as columns}

\item{distancetype}{a \code{character} that can be either \code{'pearson'}, \code{'uncentered'}, \code{'spearman'} or \code{'euclidean'}}

\item{usegpu}{\code{logical} \code{TRUE} or \code{FALSE}}
}
\value{
\code{select_distance(distancetype, usegpu)} assigns global function calculate_distance according to the parameters specified

\code{calculate_distance_pearson_gpu(x)} returns columnwise pearson distance calculated using the GPU

\code{calculate_distance_pearson_cpu(x)} returns columnwise pearson distance calculated using the CPU

\code{calculate_distance_uncentered_gpu(x)} returns columnwise uncentered pearson distance calculated using the GPU

\code{calculate_distance_uncentered_cpu(x)} returns columnwise uncentered pearson distance calculated using the CPU

\code{calculate_distance_spearman_gpu(x)} returns columnwise spearman distance calculated using the GPU

\code{calculate_distance_spearman_cpu(x)} returns columnwise spearman distance calculated using the CPU

\code{calculate_distance_euclidean_gpu(x)} returns columnwise euclidean distance calculated using the GPU

\code{calculate_distance_euclidean_cpu(x)} returns columnwise euclidean distance calculated using the CPU
}
\description{
Functions to calculate distance matrices using cpu or gpu computing
}
\examples{
rna_luad<-use_rna_luad()
prm <- rna_luad$TCGA$expression_matrix
#Will select the calculate_distance_pearson_cpu() function to calculate the distance matrix
select_distance(distancetype= "pearson", usegpu=FALSE)
calculate_distance(prm)
calculate_distance_pearson_gpu(prm)
calculate_distance_pearson_cpu(prm)
calculate_distance_uncentered_gpu(prm)
calculate_distance_uncentered_cpu(prm)
calculate_distance_spearman_gpu(prm)
calculate_distance_spearman_cpu(prm)
calculate_distance_euclidean_gpu(prm)
calculate_distance_euclidean_cpu(prm)
}
\author{
Martin E Guerrero-Gimenez, \email{mguerrero@mendoza-conicet.gob.ar}
}
