% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galgo_main.R
\name{cDist}
\alias{cDist}
\title{Restricted Mean Survival distance between groups}
\usage{
cDist(x)
}
\arguments{
\item{x}{A numeric vector of length > 1}
}
\value{
A numeric vector of length 1. The function computes the Harmonic mean of the differences between consecutive groups multiplied by the number of comparisons
This function is used inside 'fitness' function.
}
\description{
Restricted Mean Survival distance between groups
}
\examples{
V <- c(4.5,3,7,11)
cDist(V)
}
