% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_functions.R
\name{classify.multiple}
\alias{classify.multiple}
\title{Classify samples from multiple centroids}
\usage{
classify.multiple(prob_matrix, centroid.list, distancetype = "pearson")
}
\arguments{
\item{prob_matrix}{a \code{matrix} or \code{data.frame}. Must be an expression matrix with features in rows and samples in columns}

\item{centroid.list}{a\code{list} with the centroid matrix for each of the signatures to evaluate, where each column represents the prototypic centroid of a subtype and each row the constituents features of the solution signature. The output of \code{\link[galgoR:create.centroids]{create.centroids}} can be used.}

\item{distancetype}{a \code{character} that can be either \code{'pearson'} (default), \code{'spearman'} or \code{'kendall'}.}
}
\value{
Returns a \code{data.frame} with the classes assigned to each sample in each signature, were samples are a rows and signatures in columns
}
\description{
Classify samples from multiple centroids
}
\examples{
#' #Load data
require(survival)
rna_luad <- use_rna_luad()
TCGA_expr <- rna_luad$TCGA$expression_matrix
TCGA_clinic <- rna_luad$TCGA$pheno_data
OS <- survival::Surv(time=TCGA_clinic$time,event=TCGA_clinic$status)

#Run galgo
output <- galgoR::galgo(generations = 10,population = 20, prob_matrix = TCGA_expr, OS = OS)
RESULTS<- non.dominated.summary(output=output,OS=OS, prob_matrix= TCGA_expression, distancetype ="pearson", usegpu= FALSE)
CentroidsList<- create.centroids(output,RESULTS$solution,train.set= TCGA_expression)
TCGA_classes <- classify.multiple(prob_matrix=TCGA_expr,centroid.list= CentroidsList)
}
