% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{classify}
\alias{classify}
\title{Title
Given the distance to the centroids classify the samples}
\usage{
classify(data, centroid, method = "pearson")
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Title
Given the distance to the centroids classify the samples
}
