% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-functions.R
\name{classify_multiple}
\alias{classify_multiple}
\title{Classify samples from multiple centroids}
\usage{
classify_multiple(prob_matrix, centroid._list, distancetype = "pearson")
}
\arguments{
\item{prob_matrix}{a \code{matrix} or \code{data.frame}. Must be an expression matrix with features in rows and samples in columns}

\item{centroid._list}{a\code{list} with the centroid matrix for each of the signatures to evaluate, where each column represents the prototypic centroid of a subtype and each row the constituents features of the solution signature. The output of \code{\link[galgoR:create_centroids]{create_centroids}} can be used.}

\item{distancetype}{a \code{character} that can be either \code{'pearson'} (default), \code{'spearman'} or \code{'kendall'}.}
}
\value{
Returns a \code{data.frame} with the classes assigned to each sample in each signature, were samples are a rows and signatures in columns
}
\description{
Classify samples from multiple centroids
}
\examples{
\dontrun{
# Load data
rna_luad <- use_rna_luad()
TCGA_expr <- rna_luad$TCGA$expression_matrix
TCGA_clinic <- rna_luad$TCGA$pheno_data
OS <- survival::Surv(time = TCGA_clinic$time, event = TCGA_clinic$status)

# Run galgo
output <- galgoR::galgo(generations = 10, population = 30, prob_matrix = TCGA_expr, OS = OS)
RESULTS <- non_dominated_summary(
  output = output, OS = OS,
  prob_matrix = TCGA_expr,
  distancetype = "pearson",
  usegpu = FALSE
)
CentroidsList <- create_centroids(output, RESULTS$solution, trainset = TCGA_expr)
TCGA_classes <- classify_multiple(prob_matrix = TCGA_expr, centroid._list = CentroidsList)
}
}
