% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{cluster_algorithm}
\alias{cluster_algorithm}
\title{Title
To use partition around medioids (PAM), This is the default for the the current aplication}
\usage{
cluster_algorithm(c, k)
}
\arguments{
\item{k}{}
}
\value{

}
\description{
Title
To use partition around medioids (PAM), This is the default for the the current aplication
}
