% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galgo_main.R
\name{fitness}
\alias{fitness}
\title{Survival fitness function using the Restricted Mean Survival Time (RMST) of each group}
\usage{
fitness(OS, clustclass, period)
}
\arguments{
\item{OS}{a survival object with survival data of the patients evaluated}

\item{clustclass}{a numeric vector with the group label for each patient}

\item{period}{a number representing the period of time to evaluate in the RMST calculation}
}
\value{
The function computes the Harmonic mean of the differences between Restricted Mean Survival Time (RMST) of consecutive survival curves multiplied by the number of comparisons
Restricted Mean Survival Time: https://www.bmj.com/content/357/bmj.j2250
}
\description{
Survival fitness function using the Restricted Mean Survival Time (RMST) of each group
}
\examples{
rna_luad<-use_rna_luad()
clinical <- rna_luad$clinical
OS <- survival::Surv(time=clinical$time,event=clinical$status)
fitness(OS,clustclass= clinical$Wilk.Subtype, 1825)
}
