% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galgo_main.R
\name{galgo}
\alias{galgo}
\title{Galgo main function}
\usage{
galgo(
  population = 30,
  generations = 2,
  nCV = 5,
  usegpu = FALSE,
  distancetype = "pearson",
  TournamentSize = 2,
  period = 1825,
  OS,
  prob_matrix,
  res_dir = "results/",
  save_pop_partial_callback = base_save_pop_partial_callback,
  save_pop_final_callback = base_save_pop_final_callback,
  report_callback = base_report_callback,
  start_gen_callback = base_start_gen_callback,
  end_gen_callback = base_end_gen_callback,
  verbose = 2
)
}
\arguments{
\item{population}{a number indicating the number of solutions in the population of solutions that will be evolved}

\item{generations}{a number indicating the number of iterations of the galgo algorithm}

\item{nCV}{number of cross-validation sets}

\item{usegpu}{logical default FALSE, set to TRUE if you wish to use gpu computing (gpuR package must be properly installed and loaded)}

\item{distancetype}{character, it can be "pearson" (centered pearson), "uncentered" (uncentered pearson), "spearman" or "euclidean"}

\item{TournamentSize}{a number indicating the size of the tournaments for the selection procedure}

\item{period}{a number indicating the outcome period to evaluate the RMST}

\item{OS}{a survival object (see 'Surv' function from the 'survival' package)}

\item{prob_matrix}{a matrix or data.frame. Must be an expression matrix with features in rows and samples in columns}

\item{res_dir}{a character indicating where to save the intermediate and final output of the algorithm}

\item{save_pop_partial_callback}{optional callback function between iterations}

\item{save_pop_final_callback}{optional callback function for the last iteration}

\item{report_callback}{optional callback function}

\item{start_gen_callback}{optional callback function for the beginning of the run}

\item{end_gen_callback}{optional callback function for the end of the run}

\item{verbose}{?}
}
\value{
an object of type 'galgo.Obj' that corresponds to a list with the elements $Solutions and $ParetoFront. $Solutions is a l x (n + 5) matrix where n is the number of features evaluated and l is the number of solutions obtained.
The submatrix l x n is a binary matrix where each row represents the chromosome of an evolved solution from the solution population where each feature can be present (1) or absent (0). Column n +1 represent the  k number of clusters for each solutions. Column n+2 to n+5 shows the SC Fitness and Survival Fitness values, the solution rank and the crowding distance of the solution in the final pareto front respectevely.
For easier interpretation of the 'galgo.Obj', the output can be reshaped using the 'toList' and 'toDataFrame' functions
}
\description{
'galgo' accepts an expression matrix and a survival object to find robust gene expression signatures related to a given outcome
}
\examples{
rna_luad<-use_rna_luad()

prm <- rna_luad$prm
clinical <- rna_luad$clinical
OS <- survival::Surv(time=clinical$time,event=clinical$status)
output <- galgo(generations = 10, population = 30,prob_matrix = prm, OS=OS)
outputDF <- toDataFrame(output)
}
\author{
Martin E Guerrero-Gimenez, \email{mguerrero@mendoza-conicet.gob.ar}
}
