% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_functions.R
\name{non_dominated_summary}
\alias{non_dominated_summary}
\title{Summary of the non dominated solutions}
\usage{
non_dominated_summary(output, prob_matrix, OS, distancetype = "pearson",
  usegpu = FALSE)
}
\arguments{
\item{output}{An object of class \code{galgo.Obj}}

\item{prob_matrix}{a \code{matrix} or \code{data.frame}. Must be an expression matrix with features in rows and samples in columns}

\item{OS}{a \code{survival} object (see \code{\link[survival]{Surv}} function from the \code{\link{survival}} package)}

\item{distancetype}{a \code{character} that can be either \code{'pearson'}, \code{'uncentered'}, \code{'spearman'} or \code{'euclidean'}}

\item{usegpu}{\code{logical} \code{TRUE} or \code{FALSE}}
}
\value{
Returns a \code{data.frame} with 5 columns and a number of rows equals to the non-dominated solutions found by GalgoR.
The first column has the name of the non-dominated solutions, the second the number of partitions found for each solution \code{(k)}, the third, the number of genes, the fourth the mean silhouette coefficient of the solution and the last columns has the estimated C.Index for each one.
}
\description{
The function uses a \code{'galgo.Obj'} as input an the training dataset to evaluate the non-dominated solutions found by GalgoR
}
\examples{
\dontrun{
# Load data
rna_luad <- use_rna_luad()
TCGA_expr <- rna_luad$TCGA$expression_matrix
TCGA_clinic <- rna_luad$TCGA$pheno_data
OS <- survival::Surv(time = TCGA_clinic$time, event = TCGA_clinic$status)

# Run galgo
output <- galgoR::galgo(generations = 10, population = 30, prob_matrix = TCGA_expr, OS = OS)
non_dominated_summary(
  output = output, OS = OS,
  prob_matrix = TCGA_expr,
  distancetype = "pearson",
  usegpu = FALSE
)
}
}
