% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_pareto}
\alias{plot_pareto}
\title{Plot pareto front from an galgo.Obj}
\usage{
plot_pareto(output)
}
\arguments{
\item{output}{An object of class \code{galgo.Obj}}
}
\value{
This function returns a scatterplot showing the solutions found by Galgo accross all generations in the solution space, where the Silhouette Fitness is in the x-axis and the survival fitness in the y-axis.
A line is drawn over all non-dominated solutions showing the estimated Pareto front
}
\description{
Plot pareto front from an galgo.Obj
}
\examples{
\dontrun{
# Load data
rna_luad <- use_rna_luad()
TCGA_expr <- rna_luad$TCGA$expression_matrix
TCGA_clinic <- rna_luad$TCGA$pheno_data
OS <- survival::Surv(time = TCGA_clinic$time, event = TCGA_clinic$status)

# Run galgo
output <- galgoR::galgo(generations = 10, population = 30, prob_matrix = TCGA_expr, OS = OS)
plot_pareto(output)
}
}
