% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galgo_main.R
\name{toList}
\alias{toList}
\title{Convert galgo.Obj to list}
\usage{
toList(output)
}
\arguments{
\item{output}{An object of class \code{galgo.Obj}}
}
\value{
The current function restructurates a \code{galgo.Obj} to a more easy to understand an use \code{list}. This output is particularly useful if one wants to select a given solution and use its outputs in a new classifier. The output of type \code{list} has a length equals to the number of solutions obtained by the \code{\link[galgoR:galgo]{galgo}} algorithm.

Basically this output is a list of lists, where each element of the output is named after the solution's name (\code{solution.n}, where \code{n} is the number assigned to that solution), and inside of it, it has all the constituents for that given solution with the following structure:
\enumerate{
\item \strong{output$solution.n$Genes}: A vector of the features included in the solution
\item \strong{output$solution.n$k}: The number of partitions found in that solution
\item \strong{output$solution.n$SC.Fit}: The average silhouette coefficient of the partitions found
\item \strong{output$solution.n$Surv.Fit}: The survival fitness value
\item \strong{output$solution.n$Rank}: The solution rank
\item \strong{CrowD}: The solution crowding distance related to the rest of the solutions
}
}
\description{
The current function transforms a \code{galgo.Obj} to a \code{list}
}
\examples{
#Load data
rna_luad <- use_rna_luad()
TCGA_expr <- rna_luad$TCGA$expression_matrix
TCGA_clinic <- rna_luad$TCGA$pheno_data
OS <- survival::Surv(time=TCGA_clinic$time,event=TCGA_clinic$status)

#Run galgo
output <- galgoR::galgo(generations = 2 ,population = 4, prob_matrix = TCGA_expr, OS = OS)
outputList <- toList(output)
}
\author{
Martin E Guerrero-Gimenez, \email{mguerrero@mendoza-conicet.gob.ar}
}
