% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galgo_main.R
\docType{class}
\name{galgo.Obj-class}
\alias{galgo.Obj-class}
\alias{galgo.Obj}
\title{Galgo Object class}
\description{
Galgo Object class
}
\section{Slots}{

\describe{
\item{\code{Solutions}}{matrix.}

\item{\code{ParetoFront}}{list.}
}}

