% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{kcentroid}
\alias{kcentroid}
\title{Title
Function to calculate the centroids of different groups (classes)}
\usage{
kcentroid(data, class)
}
\arguments{
\item{class}{}
}
\value{

}
\description{
Title
Function to calculate the centroids of different groups (classes)
}
