% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galgo_main.R
\name{toDataFrame}
\alias{toDataFrame}
\title{Convert galgo.Obj to data.frame}
\usage{
toDataFrame(output)
}
\arguments{
\item{output}{An object of class \code{galgo.Obj}}
}
\value{
The current function restructurates a \code{galgo.Obj} to a more easy to understand an use \code{data.frame}. The output \code{data.frame} has \eqn{ m x n} dimensions, were the rownames (\eqn{m}) are the solutions obtained by the \code{\link[galgoR:galgo]{galgo}} algorithm. The columns has the following structure:
\enumerate{
 \item \strong{Genes}: The features included in each solution in form of a \code{list}
 \item \strong{k}: The number of partitions found in that solution
 \item \strong{SC.Fit}: The average silhouette coefficient of the partitions found
 \item \strong{Surv.Fit}: The survival fitness value
 \item \strong{Rank}: The solution rank
 \item \strong{CrowD}: The solution crowding distance related to the rest of the solutions
}
}
\description{
The current function transforms a \code{galgo.Obj} to a \code{data.frame}
}
\examples{
#Load data
rna_luad <- use_rna_luad()
TCGA_expr <- rna_luad$TCGA$expression_matrix
TCGA_clinic <- rna_luad$TCGA$pheno_data
OS <- survival::Surv(time=TCGA_clinic$time,event=TCGA_clinic$status)

#Run galgo
output <- galgoR::galgo(generations = 2 ,population = 5, prob_matrix = TCGA_expr, OS = OS)
outputDF <- toDataFrame(output)
}
\author{
Martin E Guerrero-Gimenez, \email{mguerrero@mendoza-conicet.gob.ar}
}
